front: move updateTrainSchedules to parseGraou helper file
This is a step in centralizing all graou parsing logic,
in order to improve code maintainability and enable us to streamline the import process.
